fix: regression on trace duration calculation
